refactor: extract forwarded public base URL helper
Move proxy-derived public base URL parsing out of index.ts into a small utility
with focused tests. This is a low-risk first slice toward breaking up the
index.ts monolith.

Refs #516.
