+++ index.ts
@@ -120,75 +120,6 @@ function decryptDownloadToken(
   }
 }
 
-function getLastHeaderValue(value: string | string[] | undefined): string | undefined {
-  const raw = Array.isArray(value) ? value[value.length - 1] : value;
-  if (!raw) return undefined;
-
-  const parts = raw
-    .split(",")
-    .map(part => part.trim())
-    .filter(Boolean);
-  return parts.length > 0 ? parts[parts.length - 1] : undefined;
-}
-
-function unquoteHeaderValue(value: string): string {
-  if (value.length >= 2 && value.startsWith('"') && value.endsWith('"')) {
-    return value.slice(1, -1).replace(/\\"/g, '"').replace(/\\\\/g, "\\");
-  }
-  return value;
-}
-
-function getForwardedPublicBaseUrl(req: Request): string | undefined {
-  if (!MCP_TRUST_PROXY) return undefined;
-
-  const forwarded = getLastHeaderValue(req.headers.forwarded);
-  const forwardedValues: Record<string, string> = {};
-
-  if (forwarded) {
-    for (const part of forwarded.split(";")) {
-      const separator = part.indexOf("=");
-      if (separator <= 0) continue;
-
-      const key = part.slice(0, separator).trim().toLowerCase();
-      const value = unquoteHeaderValue(part.slice(separator + 1).trim());
-      if (key && value) forwardedValues[key] = value;
-    }
-  }
-
-  const proto = (
-    forwardedValues.proto || getLastHeaderValue(req.headers["x-forwarded-proto"])
-  )?.toLowerCase();
-  const host = forwardedValues.host || getLastHeaderValue(req.headers["x-forwarded-host"]);
-  if (!proto || !host || !/^https?$/.test(proto)) return undefined;
-  if (/[\s/\\]/.test(host)) return undefined;
-
-  const prefix = getLastHeaderValue(req.headers["x-forwarded-prefix"]);
-  const safePrefix =
-    prefix &&
-    prefix.startsWith("/") &&
-    !prefix.startsWith("//") &&
-    !prefix.includes("://") &&
-    !/[\s\\]/.test(prefix)
-      ? prefix.replace(/\/+$/, "")
-      : undefined;
-
-  try {
-    const baseUrl = new URL(`${proto}://${host}`);
-    if (
-      baseUrl.username ||
-      baseUrl.password ||
-      baseUrl.pathname !== "/" ||
-      baseUrl.search ||
-      baseUrl.hash
-    ) {
-      return undefined;
-    }
-    if (safePrefix) baseUrl.pathname = safePrefix;
-    return baseUrl.toString().replace(/\/$/, "");
-  } catch {
-    return undefined;
-  }
-}
 
 /**
  * Build a URL pointing to the download proxy endpoint.
@@ -267,6 +198,7 @@ import { createGitLabOAuthProvider } from "./oauth-proxy.js";
 import { mcpAuthRouter } from "@modelcontextprotocol/sdk/server/auth/router.js";
 import { ipKeyGenerator } from "express-rate-limit";
 import { normalizeProxyClientIpForRateLimit } from "./utils/proxy-client-ip.js";
+import { getForwardedPublicBaseUrl } from "./utils/forwarded-public-base-url.js";
 import { normalizeGitLabApiUrl } from "./utils/url.js";
 import {
   estimateMergeCommitCount,
@@ -12474,7 +12406,7 @@ async function startStreamableHTTPServer(): Promise<void> {
       token: effective.token,
       lastUsed: Date.now(),
       apiUrl: effective.apiUrl,
-      publicBaseUrl: getForwardedPublicBaseUrl(req),
+      publicBaseUrl: getForwardedPublicBaseUrl(req, MCP_TRUST_PROXY),
     };
 
     // Step 4: create a fresh transport per request.
@@ -12701,7 +12633,7 @@ async function startStreamableHTTPServer(): Promise<void> {
   // Streamable HTTP endpoint - handles both session creation and message handling
   app.post("/mcp", mcpBearerAuth, async (req: Request, res: Response) => {
     const sessionId = readMcpSessionIdHeader(req);
-    const publicBaseUrl = getForwardedPublicBaseUrl(req);
+    const publicBaseUrl = getForwardedPublicBaseUrl(req, MCP_TRUST_PROXY);
 
     // Track request
     metrics.requestsProcessed++;
